Add multi texture support for MultiSpriteRenderer

+++ Learning2DEngine/Learning2DEngine/Render/MultiSpriteDynamicData.h
@@ -8,7 +8,7 @@ namespace Learning2DEngine
 		{
 			float modelMatrix[16];
 			float color[4];
-			int textureId;
+			float textureId;
 		};
 	}
 }
